OBMIN-205 fixed avatar at add goods
