+++ src/main/resources/react/src/components/header/nav_top/index.jsx
@@ -1,23 +1,31 @@
 import { useEffect, useState } from 'react';
-import { useSelector } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 
 import { route } from 'routes/routeConstants';
 import { EllipsisText } from 'components/common';
 import { getProfile } from 'store/profile/slice';
 import { Avatar } from 'components/common/avatar';
+import { getUserThunk } from 'store/profile/thunk';
 import { CustomSelect } from 'components/selectLang';
-import { getAuth, getAuthProfile } from 'store/auth/slice';
+import { getAuthed, getAuthProfile } from 'store/auth/slice';
 import { getTranslatedText } from 'components/local/localization';
 import { ReactComponent as HeartSvg } from 'assets/icons/heart.svg';
 
 import * as Styles from './styles';
 
 const NavTop = () => {
-  const isAuthed = useSelector(getAuth);
+  const dispatch = useDispatch();
+  const isAuthed = useSelector(getAuthed);
   const profile = useSelector(getAuthProfile);
   const { avatarImage } = useSelector(getProfile);
 
   const [image, setImage] = useState('');
+
+  useEffect(() => {
+    if (!avatarImage && isAuthed) {
+      dispatch(getUserThunk());
+    }
+  }, [dispatch, isAuthed, avatarImage]);
 
   useEffect(() => {
     if (!avatarImage?.includes('blob') && avatarImage) {
