+++ src/main/resources/react/src/pages/homepage/HomePage.jsx
@@ -1,10 +1,3 @@
-import { useEffect } from 'react';
-import { getUserThunk } from 'store/profile/thunk';
-import { useDispatch, useSelector } from 'react-redux';
-
-import { getAuthed } from 'store/auth/slice';
-import { getProfile } from 'store/profile/slice';
-
 import Sliders from './slider';
 import HeaderInMain from './headerInMain';
 import HelpChildren from './helpChildren';
@@ -13,16 +6,6 @@ import { CurrentOffers } from './currentOffers';
 import * as Styles from './styles';
 
 const HomePage = () => {
-  const dispatch = useDispatch();
-  const isAuthed = useSelector(getAuthed);
-  const { avatarImage } = useSelector(getProfile);
-
-  useEffect(() => {
-    if (!avatarImage && isAuthed) {
-      dispatch(getUserThunk());
-    }
-  }, [dispatch]);
-
   return (
     <Styles.Main>
       <HeaderInMain />
